Remove the -site option
The -site option is a misnomer. It's more of a -oasis option; all it does is to install the switch_modules script. It's inapplicable to sites in general, just to sites that are mounting OASIS/cvmfs, so we should document it differently or perhaps even install the switch_modules differently.